Split test_run into agent and controller tests
This should help clarify responsibilities doing the actual split

Fixes #806
